Initial commit
Bring things over from elm-lang/core and create a README that
introduces the concept of generators more completely.

The idea of moving it out of core is that it is a pure library that
benefits from having a separate release cycle. Longer term, it might be
nice to merge this with random-extra, in which case, the docs will be
very significantly nicer as a standalone package.