Reduce allocations in independentSeed definition
@mgold, can you verify that this change is okay?

I noticed that the old version swaps the seeds. Rather than continuing
from seed0 to seed1 (and handing out the new seed) it goes from seed0
to the new seed (and hands out seed1)

That seems weird though. Does it matter which of those it does?

Separately, why do we call `next` on the brand new seed?